Refactor wallet hooks, organize by protocol
