+++ src/components/buttons/ConnectAwareSubmitButton.tsx
@@ -4,7 +4,7 @@ import { useCallback } from 'react';
 import { ProtocolType } from '@hyperlane-xyz/utils';
 
 import { tryGetProtocolType } from '../../features/caip/chains';
-import { useAccountForChain, useConnectFns } from '../../features/wallet/hooks';
+import { useAccountForChain, useConnectFns } from '../../features/wallet/hooks/multiProtocol';
 import { useTimeout } from '../../utils/timeout';
 
 import { SolidButton } from './SolidButton';
