+++ src/features/wallet/SideBarMenu.tsx
@@ -22,7 +22,7 @@ import { getToken } from '../tokens/metadata';
 import { TransfersDetailsModal } from '../transfer/TransfersDetailsModal';
 import { TransferContext } from '../transfer/types';
 
-import { useAccounts, useDisconnectFns } from './hooks';
+import { useAccounts, useDisconnectFns } from './hooks/multiProtocol';
 
 export function SideBarMenu({
   onConnectWallet,
